+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - `formatXXX()` functions now gain a new argument `rows` (starting from 1), which can be used to specify the rows that the "style" should be applied to. Note that it only works expected in the client-side processing mode, i.e., `server = FALSE`. `formatStyle()` is the only exception that doesn't have this argument and should use `styleRow()` instead (thanks, @jrecasens @shrektan #520).
 
+- Added a new function `updateFilters` to allow users to update the filters of a DataTable object that has been updated by reference, such as with `replaceData()` (thanks, @giadasp #934, @wholmes105 #939).
+
 ## MAJOR CHANGES
 
 - Now users can provide column names of the data to `options$columnDefs$targets`. Previously, it only supports column indexes or "_all" (thanks, @shrektan #948).